requirement.txt
fastapi
uvicorn
pillow
pytesseract
transformers
torch
python-multipart